MySQL fix and single NIC

+++ file_root/mysql/init.sls
@@ -37,16 +37,49 @@ mysql_service_running:
       - ini: mysql_conf
 
 {% set root_password = salt['pillar.get']('mysql:root_password') %}
-mysql_secure_installation:
-  cmd:
-    - run
-    - name: |
+mysql_secure_installation_script:
+  file: 
+    - managed
+    - user: root
+    - group: root
+    - mode: 400
+    - name: "/tmp/mysql-secure-installation.sh"
+    - contents: |
         #!/bin/bash
-        mysqladmin -u root password "{{ root_password }}"
+        mysql -u root -p"{{ root_password }}" -e "" &> /dev/null
+        if [ $? -eq 0 ]; then
+            echo "MySQL root password was already set."
+        else
+            mysql -u root -e "" &> /dev/null
+            if [ $? -eq 0 ]; then
+                mysqladmin -u root password "{{ root_password }}"
+                echo "MySQL root password has been successfully set."
+            else
+                echo "ERROR: Cannot change MySQL root password." >&2
+                exit 1
+            fi
+        fi
         mysql -u root -p"{{ root_password }}" -e "UPDATE mysql.user SET Password=PASSWORD('{{ root_password }}') WHERE User='root';"
         mysql -u root -p"{{ root_password }}" -e "DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1');"
         mysql -u root -p"{{ root_password }}" -e "DELETE FROM mysql.user WHERE User='';"
-        mysql -u root -p"{{ root_password }}" -e "DROP DATABASE test;"
+        mysql -u root -p"{{ root_password }}" -e "use test;" &> /dev/null
+        if [ $? -eq 0 ]; then
+            mysql -u root -p"{{ root_password }}" -e "DROP DATABASE test;"
+        fi
         mysql -u root -p"{{ root_password }}" -e "FLUSH PRIVILEGES;"
+        echo "Finished MySQL secure installation."
+        exit 0
     - require:
       - service: mysql_service_running
+
+mysql_secure_installation:
+  cmd:
+    - run
+    - name: "bash /tmp/mysql-secure-installation.sh"
+    - require:
+      - file: mysql_secure_installation_script
+
+mysql_secure_installation_script_delete:
+  file:
+    - absent
+    - name: "/tmp/mysql-secure-installation.sh"
